[HelpMenu] Allow for extended help in separate widget
Allow the extended help for the current help item to be displayed
in the "description" Label widget instead of (or as well as) in the
second line of the help entry.

Use with commit 32bf28e and later in the main OpenViX repository.

(cherry picked from commit 0fff32836ef3294fa4d177a4e3aef5b47ce6478c)